create snapclient component
