+++ components/net_functions/CMakeLists.txt
@@ -1,3 +1,3 @@
 idf_component_register(SRCS "net_functions.c"
                        INCLUDE_DIRS "include"
-                       REQUIRES mdns network_interface driver)
+                       REQUIRES mdns driver)
