+++ components/snapclient/connection_handler.c
