+++ components/snapclient/connection_handler.h
