+++ components/snapclient/include/snapclient.h
@@ -1,0 +1,14 @@
+#ifndef __SNAPCLIENT_H__
+#define __SNAPCLIENT_H__
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void init_snapcast(void (*set_volume)(int), void (*set_mute)(bool));
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif // __SNAPCLIENT_H__
