+++ main/CMakeLists.txt
@@ -1,8 +1,8 @@
-idf_component_register(SRCS "main.c" "connection_handler.c"
+idf_component_register(SRCS "main.c"
                        INCLUDE_DIRS "."
 
-                       PRIV_REQUIRES esp_timer esp_wifi nvs_flash audio_board audio_hal audio_sal net_functions opus flac ota_server
-                       				 ui_http_server network_interface custom_board settings_manager tas5805m_settings
+                       PRIV_REQUIRES esp_timer nvs_flash audio_board audio_hal audio_sal net_functions ota_server lightsnapcast dsp_processor
+                       				 ui_http_server network_interface custom_board settings_manager tas5805m_settings snapclient
                        )
 
 set_source_files_properties(main.c PROPERTIES COMPILE_FLAGS -Wno-implicit-fallthrough)
