+++ main/Kconfig.projbuild
@@ -1,41 +1,5 @@
 menu "Snapclient Configuration"
-    config SNAPCLIENT_USE_TIMEFILTER
-        bool "Use Kalman Timefilter"
-        default y
-        help
-            Use Kalman Timefilter for latency to server estimation. The non default is a median filter.
-            
-    config SNAPSERVER_USE_MDNS
-        bool "Use mDNS"
-        default y
-        help
-            Use mDNS to look for the snapserver to connect to on the local network.
 
-    config SNAPCLIENT_CONNECT_IPV6
-        bool "use IPv6"
-        default n
-        help
-            Use IPv6 to connect to snapserver
-
-    config SNAPSERVER_HOST
-        string "Snapserver host"
-        default "192.168.1.158"
-        depends on !SNAPSERVER_USE_MDNS
-        help
-            Host (ip or name) of snapserver to connect to if mDNS is disabled or the mDNS resolution fails.
-
-    config SNAPSERVER_PORT
-        int "Snapserver port"
-        default 1704
-        depends on !SNAPSERVER_USE_MDNS
-        help
-            Port of the snapserver to connect to.
-
-    config SNAPCLIENT_NAME
-        string "snapclient name"
-        default "ESP32-Caster"
-        help
-            Name of the client to register the snapserver.
 
 	menu "HTTP Server Setting"
 		config WEB_PORT
@@ -44,16 +8,5 @@ menu "Snapclient Configuration"
 			help
 				HTTP server port to use.
 	endmenu
-
-	config USE_SAMPLE_INSERTION
-        bool "Use sample insertion"
-        default y
-        help
-            There are two syncronization implementations.
-
-            1. APLL tuning (ONLY use with DACs having MCLK input)
-            2. sample insertion (duplicates a sample now and then to keep up sync)
-
-            Both approaches have similar performance keeping clients in sync <= 500µs
 
 endmenu
